requirements.txt
python-telegram-bot==20.7
telethon==1.34.0
cryptg==0.4.0
aiofiles==23.2.1
aiohttp==3.9.1
python-dotenv==1.0.0